backend/requirements.txt
# FastAPI stack
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==23.0.0
python-multipart==0.0.20

# DB
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
greenlet==3.1.1

# Auth & security
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0

# Validation & settings
pydantic==2.10.3
pydantic-settings==2.6.1
eval-type-backport==0.2.2
email-validator==2.2.0

# Rate limiting
slowapi==0.1.9

# Utils
python-dateutil==2.9.0.post0

# Migrations (sync driver)
psycopg[binary]==3.2.3
